+++ Assets/Scripts/Character/Components/Live/ILiveComponent.cs
@@ -8,9 +8,9 @@ namespace OmniumLessons
         public event Action<Character> OnCharacterHealthChange;
 
         public bool IsAlive { get; }
-		public int MaxHealth { get; }
+		public float MaxHealth { get; }
 		public float Health { get; }
 		
-		public void GetDamage(int damage);
+		public void GetDamage(float damage);
 	}
 }
